8 byte correction in malloc in index_nsg.h

+++ src/index_nsg.cpp
@@ -234,35 +234,35 @@ namespace efanna2e {
   void IndexNSG::init_graph_bf(const Parameters &parameters) {
     // allocate and init centroid
     float *center = new float[dimension_]();
-    for (unsigned j = 0; j < dimension_; j++)
+    for (size_t j = 0; j < dimension_; j++)
       center[j] = 0;
-    for (unsigned i = 0; i < nd_; i++) {
-      for (unsigned j = 0; j < dimension_; j++) {
+    for (size_t i = 0; i < nd_; i++) {
+      for (size_t j = 0; j < dimension_; j++) {
 	center[j] += data_[i * dimension_ + j];
       }
     }
-    for (unsigned j = 0; j < dimension_; j++) {
+    for (size_t j = 0; j < dimension_; j++) {
       center[j] /= nd_;
     }
     
     // compute all to one distance
     float* distances = new float[nd_]();
 #pragma omp parallel for schedule(static, 65536)
-    for(unsigned i=0;i<nd_;i++){
+    for(size_t i=0;i<nd_;i++){
       // extract point and distance reference
       float &dist = distances[i];
-      const float* cur_vec = data_ + (i * dimension_);
+      const float* cur_vec = data_ + ((size_t)i * (size_t)dimension_);
       dist = 0;
       float diff = 0;
-      for(unsigned j=0;j<dimension_;j++){
+      for(size_t j=0;j<dimension_;j++){
 	diff = (center[j] - cur_vec[j]) *(center[j] - cur_vec[j]);
 	dist += diff;
       }
     }
     // find imin
-    unsigned min_idx = 0;
+    size_t min_idx = 0;
     float min_dist = distances[0];
-    for(unsigned i=1;i<nd_;i++){
+    for(size_t i=1;i<nd_;i++){
       if (distances[i] < min_dist){
 	min_idx = i;
 	min_dist = distances[i];
